+++ example/src/app.tsx
@@ -1,7 +1,8 @@
 import { DEBUGGER } from 'woby'
-import * as sample from './sample.test'
-import * as comps from './components.test'
-import { Chk, Csf } from 'chk'
+import { MyComponent } from './components/components.test'
+// import * as sample from './sample.test'
+// import * as comps from './components.test'
+// import { Chk, Csf } from 'chk'
 // import * as example from './example.test'
 
 interface MyComponentProps {
@@ -10,16 +11,6 @@ interface MyComponentProps {
     timestamp: Date
 }
 
-function MyComponent(props: MyComponentProps) {
-    return (
-        <div>
-            <h1>Hello, {props.message}!</h1>
-            <p>Count: {props.count}</p>
-            <p>Timestamp: {props.timestamp.toISOString()}</p>
-            <button>Click Me</button>
-        </div>
-    )
-}
 
 const isDev = typeof import.meta.env !== 'undefined' && import.meta.env.DEV
 
@@ -32,20 +23,15 @@ export function App() {
     console.log('APP')
     return (
         <div>
-            Explicit test
-            <Chk name="MyComponent 1" message="World" count={1} timestamp={now}>
-                <MyComponent message="World" count={1} timestamp={now} />
-            </Chk>
+            Real page, no test
+            <MyComponent class="border m-10" message="World" count={1} timestamp={now} />
+            <MyComponent class="border m-10" message="Woby" count={2} timestamp={new Date(new Date(2025, 0, 2, 0, 0, 0, 0))} />
 
-            <Chk name="MyComponent 2" message="Woby" count={2} timestamp={new Date(2025, 0, 2, 0, 0, 0, 0)}>
-                <MyComponent message="Woby" count={2} timestamp={new Date(new Date(2025, 0, 2, 0, 0, 0, 0))} />
-            </Chk>
-
-            Components Story Format
+            {/* Components Story Format
             <Csf module={comps} path='components.test' />
 
             Run test(), see browser debugger console
-            <Csf module={sample} />
+            <Csf module={sample} /> */}
 
         </div>
     )