+++ example/src/components/components.test.tsx
@@ -1,0 +1,3 @@
+// Import the components from their individual files
+export * from './MyComponent'
+export * from './MyComponent2'