+++ example/src/components/index.ts
@@ -1,0 +1,3 @@
+export { MyComponent } from './MyComponent'
+export { AnotherComponent } from './AnotherComponent'
+export { MyComponent2 } from './MyComponent2'